Move add_libs_into_aapt2_test to tests folder
